// frontend/lib/constants/enums.ts
export enum Routes {
  HOME = "/",
  LOGIN = "/login",
  REGISTER = "/register",
  ADMIN = "/admin",
  LESSONS = "/lessons",
  PROFILE = "/profile",
  SETTINGS = "/settings",
  LEARN_MORE = "/learn-more",
}

export enum Roles {
  ADMIN = "ADMIN",
  USER = "USER",
}

export enum LessonLevel {
  BEGINNER = "BEGINNER",
  INTERMEDIATE = "INTERMEDIATE",
  ADVANCED = "ADVANCED",
}
export enum LessonStatus {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
}